+++ src/services/axios/get/getSensitivityCalculator/sensitivityCalculator/getSensitivityCalculatorAPIData.ts
@@ -1,4 +1,3 @@
-import { TELESCOPE_LOW, TELESCOPE_MID } from '@ska-telescope/ska-gui-components';
 import { Telescope } from '@ska-telescope/ska-gui-local-storage';
 import Observation from '../../../../../utils/types/observation';
 import Target from '../../../../../utils/types/target';
@@ -17,6 +16,7 @@ import {
 import GetZoomData from '../getZoomData/getZoomData';
 import GetContinuumData from '../getContinuumData/getContinuumData';
 import { DataProductSDPNew, SDPSpectralData } from '@/utils/types/dataProduct';
+import { getTelescope } from '@utils/helpers.ts';
 
 async function getSensCalc(
   observation: Observation,
@@ -54,8 +54,5 @@ async function getSensCalc(
       };
   }
 }
-
-const getTelescope = (telNum: number): Telescope =>
-  telNum === TELESCOPE_LOW_NUM ? TELESCOPE_LOW : TELESCOPE_MID;
 
 export default getSensCalc;