+++ src/utils/helpers.ts
@@ -14,12 +14,15 @@ import {
   TIME_UNITS,
   FREQUENCY_STR_KHZ,
   FREQUENCY_STR_MHZ,
-  FREQUENCY_STR_HZ
+  FREQUENCY_STR_HZ,
+  TELESCOPE_LOW_NUM
 } from './constants';
 import Observation from './types/observation';
 import { ValueUnitPair } from './types/valueUnitPair';
 import { OSD_CONSTANTS } from './OSDConstants';
 import { channelsToBandwidthHz } from '@utils/zoomWindow.ts';
+import { Telescope } from '@ska-telescope/ska-gui-local-storage';
+import { TELESCOPE_LOW, TELESCOPE_MID } from '@ska-telescope/ska-gui-components';
 
 export const arraysAreEqual = (a: any[], b: any[]) => {
   if (a === b) return true;
@@ -305,3 +308,6 @@ export const convertFrequencyToDisplayUnits = (
 
   return { value, unit };
 };
+
+export const getTelescope = (telNum: number): Telescope =>
+  telNum === TELESCOPE_LOW_NUM ? TELESCOPE_LOW : TELESCOPE_MID;